Former-commit-id: 857ddbf35e98a6724f0ded57539dfde2b2b2dbc2

+++ cocos2d/Platforms/iOS/CCDirectorIOS.m
@@ -344,7 +344,7 @@
 
 -(void)getFPSImageData:(unsigned char**)datapointer length:(NSUInteger*)len contentScale:(CGFloat *)scale
 {
-	int device = [[CCConfiguration sharedConfiguration] runningDevice];
+	NSInteger device = [[CCConfiguration sharedConfiguration] runningDevice];
 
 	if( device == CCDeviceiPadRetinaDisplay) {
 		*datapointer = cc_fps_images_ipadhd_png;
